Reanmed reducers
...for more clarity. Also extracted the initReducer into its own file, jsut for
coherence.
