+++ src/redux/store.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import { createLogger } from 'redux-logger';
 
-import rootReducer from './reducers'
+import rootReducer from './rootReducer';
 
 
 const loggerMiddleware = createLogger();
